🐛 메타데이터 적용을 위해 tomong/page.js 내용을 별도 컴포넌트로 분리
