+++ src/components/tomong/Result.module.css
